+++ changelog.md
@@ -1,9 +1,16 @@
 Unrealised
 ----------
 
+1.10.0
+------
+* Make Okovision International by adding English language (default).
+  * For changing Language, go to parameters menu
+* This is the last update for the coming months!  
+
 1.9.2
 ------
 * Fix #108 - Error Message rendu.getIndicByDay when you don' have pumpe Hot water
+
 
 1.9.1
 ------